fix some bugs, add Grid 70%, still lagging, intro not running properly

+++ 05-ScenceManager/Boomerang.h
@@ -10,7 +10,7 @@ public:
 	void GetBoundingBox(float& left, float& top, float& right, float& bottom);
 	void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects);
 	void Render();
-	Boomerang(float X, float Y, int dir);
+	Boomerang(float X, float Y, float dir);
 	~Boomerang();
 };
